Fold TeiMergerScenarioTest into the main MergerScenarioTest
